+++ contracts/token/oracles/IOracleAggregator.sol
@@ -4,5 +4,5 @@ pragma solidity ^0.8.20;
 interface IOracleAggregator {
     function getTokenValueOfOneNativeToken(
         address _token
-    ) external view returns (uint256 exchangeRate);
+    ) external view returns (uint128 exchangeRate);
 }
