+++ test/foundry/token-paymaster/TokenPaymaster.t.sol
@@ -195,7 +195,7 @@ contract TokenPaymasterTest is SATestBase {
                 .ORACLE_BASED;
         uint48 validUntil = 3735928559;
         uint48 validAfter = 4660;
-        uint256 exchangeRate = 977100;
+        uint128 exchangeRate = 977100;
         uint32 priceMarkup = 1100000;
 
         op.paymasterAndData = abi.encodePacked(
@@ -302,7 +302,7 @@ contract TokenPaymasterTest is SATestBase {
                 .ORACLE_BASED;
         uint48 validUntil = 3735928559;
         uint48 validAfter = 4660;
-        uint256 exchangeRate = 977100;
+        uint128 exchangeRate = 977100;
         uint32 priceMarkup = 1100000;
 
         op.paymasterAndData = abi.encodePacked(
@@ -352,7 +352,7 @@ contract TokenPaymasterTest is SATestBase {
                 .ORACLE_BASED;
         uint48 validUntil = 3735928559;
         uint48 validAfter = 4660;
-        uint256 exchangeRate = 977100;
+        uint128 exchangeRate = 977100;
         uint32 priceMarkup = 1100000;
 
         op.paymasterAndData = abi.encodePacked(
@@ -397,7 +397,7 @@ contract TokenPaymasterTest is SATestBase {
                 .ORACLE_BASED;
         uint48 validUntil = 3735928559;
         uint48 validAfter = 4660;
-        uint256 exchangeRate = 977100;
+        uint128 exchangeRate = 977100;
         uint32 priceMarkup = 2200000;
 
         bytes32 hash = _btpm.getHash(
@@ -451,7 +451,7 @@ contract TokenPaymasterTest is SATestBase {
                 .ORACLE_BASED;
         uint48 validUntil = 3735928559;
         uint48 validAfter = 4660;
-        uint256 exchangeRate = 977100;
+        uint128 exchangeRate = 977100;
         uint32 priceMarkup = 1100000;
 
         bytes32 hash = _btpm.getHash(
